change networkpolicyInformation to augmented obejct

+++ sensor/kubernetes/listener/resources/networkpolicy.go
@@ -3,6 +3,7 @@ package resources
 import (
 	"github.com/stackrox/rox/generated/internalapi/central"
 	"github.com/stackrox/rox/generated/storage"
+	"github.com/stackrox/rox/pkg/features"
 	networkPolicyConversion "github.com/stackrox/rox/pkg/protoconv/networkpolicy"
 	"github.com/stackrox/rox/sensor/common/detector"
 	"github.com/stackrox/rox/sensor/common/store"
@@ -32,22 +33,45 @@ func (h *networkPolicyDispatcher) ProcessEvent(obj, _ interface{}, action centra
 
 	roxNetpol := networkPolicyConversion.KubernetesNetworkPolicyWrap{NetworkPolicy: np}.ToRoxNetworkPolicy()
 
-	// TODO: feature flag
+	if features.NetworkPolicySystemPolicy.Enabled() {
 
-	//netpolWrap := &networkPolicyWrap{
-	//	NetworkPolicy: roxNetpol,
-	//	selector:      SelectorFromMap(np.Spec.PodSelector.MatchLabels),
-	//}
+		//netpolWrap := &networkPolicyWrap{
+		//	NetworkPolicy: roxNetpol,
+		//	selector:      SelectorFromMap(np.Spec.PodSelector.MatchLabels),
+		//}
 
-	var sel selector
-	oldWrap := h.netpolStore.Get(roxNetpol.GetId())
-	if oldWrap != nil {
-		sel = SelectorFromMap(oldWrap.GetSpec().GetPodSelector().GetMatchLabels())
-	}
+		var sel selector
+		oldWrap := h.netpolStore.Get(roxNetpol.GetId())
+		if oldWrap != nil {
+			sel = SelectorFromMap(oldWrap.GetSpec().GetPodSelector().GetMatchLabels())
+		}
 
-	if action == central.ResourceAction_REMOVE_RESOURCE {
-		h.netpolStore.Delete(roxNetpol.GetId(), roxNetpol.GetNamespace())
+		if action == central.ResourceAction_REMOVE_RESOURCE {
+			h.netpolStore.Delete(roxNetpol.GetId(), roxNetpol.GetNamespace())
+			h.updateDeploymentsFromStore(roxNetpol, sel, action)
+			return []*central.SensorEvent{
+				{
+					Id:     string(np.UID),
+					Action: action,
+					Resource: &central.SensorEvent_NetworkPolicy{
+						NetworkPolicy: roxNetpol,
+					},
+				},
+			}
+		}
+
+		h.netpolStore.Upsert(roxNetpol)
+		if action == central.ResourceAction_UPDATE_RESOURCE {
+			if sel != nil {
+				sel = or(sel, SelectorFromMap(roxNetpol.GetSpec().GetPodSelector().GetMatchLabels()))
+			} else {
+				sel = SelectorFromMap(roxNetpol.GetSpec().GetPodSelector().GetMatchLabels())
+			}
+		} else if action == central.ResourceAction_CREATE_RESOURCE {
+			sel = SelectorFromMap(roxNetpol.GetSpec().GetPodSelector().GetMatchLabels())
+		}
 		h.updateDeploymentsFromStore(roxNetpol, sel, action)
+
 		return []*central.SensorEvent{
 			{
 				Id:     string(np.UID),
@@ -58,18 +82,6 @@ func (h *networkPolicyDispatcher) ProcessEvent(obj, _ interface{}, action centra
 			},
 		}
 	}
-
-	h.netpolStore.Upsert(roxNetpol)
-	if action == central.ResourceAction_UPDATE_RESOURCE {
-		if sel != nil {
-			sel = or(sel, SelectorFromMap(roxNetpol.GetSpec().GetPodSelector().GetMatchLabels()))
-		} else {
-			sel = SelectorFromMap(roxNetpol.GetSpec().GetPodSelector().GetMatchLabels())
-		}
-	} else if action == central.ResourceAction_CREATE_RESOURCE {
-		sel = SelectorFromMap(roxNetpol.GetSpec().GetPodSelector().GetMatchLabels())
-	}
-	h.updateDeploymentsFromStore(roxNetpol, sel, action)
 
 	return []*central.SensorEvent{
 		{
